Refactor neptune memory to include a neptune base
